Fix unstable boundary (fixed boundary)
Unstable particles at boundary is because of the integration. If the
time_step is too big, distance of particles may be approximated to zero.
Therefore, the value scaling factor will be infinite.
To fix this, I choose to use smaller timestep instead in this commit.
In the future updates, I will use static boundary particles to resolve this issue.